+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Added `Automatic-Module-Name` to JAR files
 - Added support for Linux ARM. Compiled on a Raspberry Pi 2.
 - Added support for Linux ARM-64. Compiled on a Raspberry Pi 3.
+- Added `Argon2Helper` class with a method to find the optimal number of iterations.
 
 ### Fixed
 - Fixed wrong Argon2 id mapping ([PR](https://github.com/phxql/argon2-jvm/pull/41))